// eic-work-permit-admin/src/common/enums.ts
export enum Status {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}

export enum Gender {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum CurrencyTypes {
  BIRR = "BIRR",
  DOLLAR = "DOLLAR"
}

export enum Units {
  SQUARE_METERS = "SQUARE_METERS",
}

export enum IRoleAccount {
  INSPECTOR = "INSPECTOR",
  CASE_WORKER = "CASE_WORKER",
  TEAM_LEADER = "TEAM_LEADER",
  DIRECTOR = "DIRECTOR"
}

export enum InvestmentPermitStatus {
  DRAFTED = "DRAFTED",
  ACCEPTED = "ACCEPTED",
  REVIEWED = "REVIEWED",
  VERIFIED = "VERIFIED",
  APPROVED = "APPROVED",
  RENEWED = "RENEWED",
  CANCELLED = "CANCELLED",
  SENT_COMPANY_NAME = "SENT_COMPANY_NAME",
  APPROVED_COMPANY_NAME = "APPROVED_COMPANY_NAME",
  SENT_NEW_COMPANY_NAME = "SENT_NEW_COMPANY_NAME",
  REGISTERED_COMPANY_NAME = "REGISTERED_COMPANY_NAME",
  SENT_BANK_SLIP = "SENT_BANK_SLIP",
  ACCEPTED_BANK_SLIP = "ACCEPTED_BANK_SLIP"
}